add explicit casts to remove compiler warnings

+++ apps/NSXQt/dialogs/LogFileDialog.cpp
@@ -86,7 +86,7 @@ double LogFileDialog::dmax() const
     return ui->dMaxBox->value();
 }
 
-double LogFileDialog::numShells() const
+int LogFileDialog::numShells() const
 {
     return ui->shellBox->value();
 }
